SBERDOMA-62 extract BILLING_ACCOUNT_METER_FIELD and PERIOD_FIELD
