+++ apps/condo/schema/Billing.js
@@ -3,30 +3,13 @@ const { Text, Relationship, Integer, DateTimeUtc, CalendarDay } = require('@keys
 const { GQLListSchema } = require('@core/keystone/schema')
 const { Json } = require('@core/keystone/fields')
 const { historical, versioned, uuided, tracked, softDeleted } = require('@core/keystone/plugins')
-const { INTEGRATION_CONTEXT_FIELD, IMPORT_ID_FIELD, RAW_DATA_FIELD, BILLING_PROPERTY_FIELD, BILLING_ACCOUNT_FIELD } = require('@condo/domains/billing/schema/fields')
+const { INTEGRATION_CONTEXT_FIELD, IMPORT_ID_FIELD, RAW_DATA_FIELD, BILLING_PROPERTY_FIELD, BILLING_ACCOUNT_FIELD, BILLING_ACCOUNT_METER_FIELD, PERIOD_FIELD } = require('@condo/domains/billing/schema/fields')
 
 const { SENDER_FIELD, DV_FIELD, ORGANIZATION_OWNED_FIELD } = require('./_common')
 const { WRONG_TEXT_FORMAT, JSON_EXPECT_OBJECT_ERROR, DV_UNKNOWN_VERSION_ERROR, JSON_UNKNOWN_VERSION_ERROR } = require('../constants/errors')
 const { UPPER_CASE_ALPHANUMERIC_REGEXP } = require('../constants/regexps')
 const { rules } = require('../access')
 const { hasRequestAndDbFields, hasValidJsonStructure } = require('../utils/validation.utils')
-
-
-const BILLING_ACCOUNT_METER_FIELD = {
-    schemaDoc: 'Billing account meter',
-    type: Relationship,
-    ref: 'BillingAccountMeter',
-    isRequired: true,
-    knexOptions: { isNotNullable: true }, // Relationship only!
-    kmigratorOptions: { null: false, on_delete: 'models.CASCADE' },
-}
-
-const PERIOD_FIELD = {
-    schemaDoc: 'Period date (01.2020, 02.2020, ...)',
-    type: CalendarDay,
-    isRequired: true,
-    // TODO(pahaz): validate it
-}
 
 // 0. CHECK IS INTEGRATION EXISTS AND IS INTEGRATION TOKEN CORRECT
 // 1. GET ALL INTEGRATION ORGANIZATION CONTEXTS (property: null, unit: null)
